+++ ydb/core/tx/replication/controller/secret_resolver.cpp
@@ -3,7 +3,7 @@
 #include "secret_resolver.h"
 
 #include <ydb/core/kqp/common/events/script_executions.h>
-#include <ydb/core/kqp/federated_query/actors/kqp_federated_query_actors.h>
+#include <ydb/services/secret/resolver.h>
 #include <ydb/core/tx/scheme_cache/scheme_cache.h>
 #include <ydb/library/actors/core/actor_bootstrapped.h>
 #include <ydb/library/actors/core/hfunc.h>
@@ -46,17 +46,17 @@ class TSecretResolver: public TActorBootstrapped<TSecretResolver> {
         const auto& userSID = entry.SecurityObject->GetOwnerSID();
         SecretId = NMetadata::NSecret::TSecretId(userSID, SecretName);
 
-        if (NKqp::UseSchemaSecrets(AppData()->FeatureFlags, SecretId.GetSecretId())) {
+        if (NSecret::UseSchemaSecrets(AppData()->FeatureFlags, SecretId.GetSecretId())) {
             const TVector<TString> secretNames{SecretId.GetSecretId()};
             auto userToken = MakeIntrusiveConst<NACLib::TUserToken>(userSID, TVector<TString>());
             const auto actorSystem = ActorContext().ActorSystem();
             const auto replyActorId = SelfId();
-            auto future = NKqp::DescribeSecret(
+            auto future = NSecret::DescribeSecret(
                 secretNames,
                 userToken,
                 Database,
                 actorSystem,
-                {.RetryPolicy = NKqp::MakeLongRetryPolicy()});
+                {.RetryPolicy = NSecret::MakeLongRetryPolicy()});
             future.Subscribe([actorSystem, replyActorId](const NThreading::TFuture<NKqp::TEvDescribeSecretsResponse::TDescription>& result) {
                 actorSystem->Send(replyActorId, new NKqp::TEvDescribeSecretsResponse(result.GetValue()));
             });
