+++ ydb/services/secret/ut/common/ya.make
@@ -8,8 +8,8 @@ PEERDIR(
     ydb/public/sdk/cpp/src/client/table
     ydb/library/aclib
     ydb/core/kqp/common/events
-    ydb/core/kqp/federated_query/actors
     ydb/core/kqp/ut/common
+    ydb/services/secret
 )
 
 YQL_LAST_ABI_VERSION()
